@@ -1,23 +1,24 @@
 name: chewie
 description: A video player for Flutter with Cupertino and Material play controls
-version: 0.12.2
-homepage: https://github.com/brianegan/chewie
+version: 1.5.0
+homepage: https://github.com/fluttercommunity/chewie
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.22.0 <2.0.0"
+  sdk: '>=2.13.0 <4.0.0'
+  flutter: '>=1.22.0'
 
 dependencies:
-  cupertino_icons: ^1.0.0
+  cupertino_icons: ^1.0.5
   flutter:
     sdk: flutter
-  video_player: ^1.0.0
-  wakelock: ^0.2.1+1
+  provider: ^6.0.3
+  video_player: ^2.4.7
+  wakelock: ^0.6.2
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  lint: ^1.5.1
+  flutter_lints: ^2.0.1
 
 flutter:
   uses-material-design: true
