@@ -1,10 +1,14 @@
 import 'dart:async';
 
 import 'package:chewie/src/chewie_progress_colors.dart';
+import 'package:chewie/src/models/option_item.dart';
+import 'package:chewie/src/models/options_translation.dart';
+import 'package:chewie/src/models/subtitle_model.dart';
+import 'package:chewie/src/notifiers/player_notifier.dart';
 import 'package:chewie/src/player_with_controls.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter/widgets.dart';
+import 'package:provider/provider.dart';
 import 'package:video_player/video_player.dart';
 import 'package:wakelock/wakelock.dart';
 
@@ -12,7 +16,7 @@ typedef ChewieRoutePageBuilder = Widget Function(
   BuildContext context,
   Animation<double> animation,
   Animation<double> secondaryAnimation,
-  _ChewieControllerProvider controllerProvider,
+  ChewieControllerProvider controllerProvider,
 );
 
 /// A Video Player with Material and Cupertino skins.
@@ -21,10 +25,9 @@ typedef ChewieRoutePageBuilder = Widget Function(
 /// make it easy to use!
 class Chewie extends StatefulWidget {
   const Chewie({
-    Key key,
-    @required this.controller,
-  })  : assert(controller != null, 'You must provide a chewie controller'),
-        super(key: key);
+    Key? key,
+    required this.controller,
+  }) : super(key: key);
 
   /// The [ChewieController]
   final ChewieController controller;
@@ -38,10 +41,14 @@ class Chewie extends StatefulWidget {
 class ChewieState extends State<Chewie> {
   bool _isFullScreen = false;
 
+  bool get isControllerFullScreen => widget.controller.isFullScreen;
+  late PlayerNotifier notifier;
+
   @override
   void initState() {
     super.initState();
     widget.controller.addListener(listener);
+    notifier = PlayerNotifier.init();
   }
 
   @override
@@ -56,30 +63,39 @@ class ChewieState extends State<Chewie> {
       widget.controller.addListener(listener);
     }
     super.didUpdateWidget(oldWidget);
+    if (_isFullScreen != isControllerFullScreen) {
+      widget.controller._isFullScreen = _isFullScreen;
+    }
   }
 
   Future<void> listener() async {
-    if (widget.controller.isFullScreen && !_isFullScreen) {
-      _isFullScreen = true;
+    if (isControllerFullScreen && !_isFullScreen) {
+      _isFullScreen = isControllerFullScreen;
       await _pushFullScreenWidget(context);
     } else if (_isFullScreen) {
-      Navigator.of(context, rootNavigator: true).pop();
+      Navigator.of(
+        context,
+        rootNavigator: widget.controller.useRootNavigator,
+      ).pop();
       _isFullScreen = false;
     }
   }
 
   @override
   Widget build(BuildContext context) {
-    return _ChewieControllerProvider(
+    return ChewieControllerProvider(
       controller: widget.controller,
-      child: const PlayerWithControls(),
+      child: ChangeNotifierProvider<PlayerNotifier>.value(
+        value: notifier,
+        builder: (context, w) => const PlayerWithControls(),
+      ),
     );
   }
 
   Widget _buildFullScreenVideo(
     BuildContext context,
     Animation<double> animation,
-    _ChewieControllerProvider controllerProvider,
+    ChewieControllerProvider controllerProvider,
   ) {
     return Scaffold(
       resizeToAvoidBottomInset: false,
@@ -95,11 +111,11 @@ class ChewieState extends State<Chewie> {
     BuildContext context,
     Animation<double> animation,
     Animation<double> secondaryAnimation,
-    _ChewieControllerProvider controllerProvider,
+    ChewieControllerProvider controllerProvider,
   ) {
     return AnimatedBuilder(
       animation: animation,
-      builder: (BuildContext context, Widget child) {
+      builder: (BuildContext context, Widget? child) {
         return _buildFullScreenVideo(context, animation, controllerProvider);
       },
     );
@@ -110,17 +126,28 @@ class ChewieState extends State<Chewie> {
     Animation<double> animation,
     Animation<double> secondaryAnimation,
   ) {
-    final controllerProvider = _ChewieControllerProvider(
+    final controllerProvider = ChewieControllerProvider(
       controller: widget.controller,
-      child: const PlayerWithControls(),
+      child: ChangeNotifierProvider<PlayerNotifier>.value(
+        value: notifier,
+        builder: (context, w) => const PlayerWithControls(),
+      ),
     );
 
     if (widget.controller.routePageBuilder == null) {
       return _defaultRoutePageBuilder(
-          context, animation, secondaryAnimation, controllerProvider);
+        context,
+        animation,
+        secondaryAnimation,
+        controllerProvider,
+      );
     }
-    return widget.controller.routePageBuilder(
-        context, animation, secondaryAnimation, controllerProvider);
+    return widget.controller.routePageBuilder!(
+      context,
+      animation,
+      secondaryAnimation,
+      controllerProvider,
+    );
   }
 
   Future<dynamic> _pushFullScreenWidget(BuildContext context) async {
@@ -135,7 +162,10 @@ class ChewieState extends State<Chewie> {
       Wakelock.enable();
     }
 
-    await Navigator.of(context, rootNavigator: true).push(route);
+    await Navigator.of(
+      context,
+      rootNavigator: widget.controller.useRootNavigator,
+    ).push(route);
     _isFullScreen = false;
     widget.controller.exitFullScreen();
 
@@ -143,10 +173,13 @@ class ChewieState extends State<Chewie> {
     // so we do not need to check Wakelock.isEnabled.
     Wakelock.disable();
 
-    SystemChrome.setEnabledSystemUIOverlays(
-        widget.controller.systemOverlaysAfterFullScreen);
+    SystemChrome.setEnabledSystemUIMode(
+      SystemUiMode.manual,
+      overlays: widget.controller.systemOverlaysAfterFullScreen,
+    );
     SystemChrome.setPreferredOrientations(
-        widget.controller.deviceOrientationsAfterFullScreen);
+      widget.controller.deviceOrientationsAfterFullScreen,
+    );
   }
 
   void onEnterFullScreen() {
@@ -154,19 +187,24 @@ class ChewieState extends State<Chewie> {
     final videoHeight =
         widget.controller.videoPlayerController.value.size.height;
 
-    if (widget.controller.systemOverlaysOnEnterFullScreen != null) {
-      /// Optional user preferred settings
-      SystemChrome.setEnabledSystemUIOverlays(
-          widget.controller.systemOverlaysOnEnterFullScreen);
-    } else {
-      /// Default behavior
-      SystemChrome.setEnabledSystemUIOverlays([]);
-    }
+    SystemChrome.setEnabledSystemUIMode(SystemUiMode.manual, overlays: []);
+
+    // if (widget.controller.systemOverlaysOnEnterFullScreen != null) {
+    //   /// Optional user preferred settings
+    //   SystemChrome.setEnabledSystemUIMode(
+    //     SystemUiMode.manual,
+    //     overlays: widget.controller.systemOverlaysOnEnterFullScreen,
+    //   );
+    // } else {
+    //   /// Default behavior
+    //   SystemChrome.setEnabledSystemUIMode(SystemUiMode.manual, overlays: SystemUiOverlay.values);
+    // }
 
     if (widget.controller.deviceOrientationsOnEnterFullScreen != null) {
       /// Optional user preferred settings
       SystemChrome.setPreferredOrientations(
-          widget.controller.deviceOrientationsOnEnterFullScreen);
+        widget.controller.deviceOrientationsOnEnterFullScreen!,
+      );
     } else {
       final isLandscapeVideo = videoWidth > videoHeight;
       final isPortraitVideo = videoWidth < videoHeight;
@@ -208,10 +246,12 @@ class ChewieState extends State<Chewie> {
 /// `VideoPlayerController`.
 class ChewieController extends ChangeNotifier {
   ChewieController({
-    @required this.videoPlayerController,
+    required this.videoPlayerController,
+    this.optionsTranslation,
     this.aspectRatio,
     this.autoInitialize = false,
     this.autoPlay = false,
+    this.draggableProgressBar = true,
     this.startAt,
     this.looping = false,
     this.fullScreenByDefault = false,
@@ -220,7 +260,15 @@ class ChewieController extends ChangeNotifier {
     this.placeholder,
     this.overlay,
     this.showControlsOnInitialize = true,
+    this.showOptions = true,
+    this.optionsBuilder,
+    this.additionalOptions,
     this.showControls = true,
+    this.transformationController,
+    this.zoomAndPan = false,
+    this.maxScale = 2.5,
+    this.subtitle,
+    this.subtitleBuilder,
     this.customControls,
     this.errorBuilder,
     this.allowedScreenSleep = true,
@@ -228,18 +276,154 @@ class ChewieController extends ChangeNotifier {
     this.allowFullScreen = true,
     this.allowMuting = true,
     this.allowPlaybackSpeedChanging = true,
+    this.useRootNavigator = true,
     this.playbackSpeeds = const [0.25, 0.5, 0.75, 1, 1.25, 1.5, 1.75, 2],
     this.systemOverlaysOnEnterFullScreen,
     this.deviceOrientationsOnEnterFullScreen,
     this.systemOverlaysAfterFullScreen = SystemUiOverlay.values,
     this.deviceOrientationsAfterFullScreen = DeviceOrientation.values,
     this.routePageBuilder,
-  })  : assert(videoPlayerController != null,
-            'You must provide a controller to play a video'),
-        assert(playbackSpeeds.every((speed) => speed > 0),
-            'The playbackSpeeds values must all be greater than 0') {
+    this.progressIndicatorDelay,
+    this.hideControlsTimer = defaultHideControlsTimer,
+    this.controlsSafeAreaMinimum = EdgeInsets.zero,
+  }) : assert(
+          playbackSpeeds.every((speed) => speed > 0),
+          'The playbackSpeeds values must all be greater than 0',
+        ) {
     _initialize();
   }
+
+  ChewieController copyWith({
+    VideoPlayerController? videoPlayerController,
+    OptionsTranslation? optionsTranslation,
+    double? aspectRatio,
+    bool? autoInitialize,
+    bool? autoPlay,
+    bool? draggableProgressBar,
+    Duration? startAt,
+    bool? looping,
+    bool? fullScreenByDefault,
+    ChewieProgressColors? cupertinoProgressColors,
+    ChewieProgressColors? materialProgressColors,
+    Widget? placeholder,
+    Widget? overlay,
+    bool? showControlsOnInitialize,
+    bool? showOptions,
+    Future<void> Function(BuildContext, List<OptionItem>)? optionsBuilder,
+    List<OptionItem> Function(BuildContext)? additionalOptions,
+    bool? showControls,
+    TransformationController? transformationController,
+    bool? zoomAndPan,
+    double? maxScale,
+    Subtitles? subtitle,
+    Widget Function(BuildContext, dynamic)? subtitleBuilder,
+    Widget? customControls,
+    Widget Function(BuildContext, String)? errorBuilder,
+    bool? allowedScreenSleep,
+    bool? isLive,
+    bool? allowFullScreen,
+    bool? allowMuting,
+    bool? allowPlaybackSpeedChanging,
+    bool? useRootNavigator,
+    Duration? hideControlsTimer,
+    EdgeInsets? controlsSafeAreaMinimum,
+    List<double>? playbackSpeeds,
+    List<SystemUiOverlay>? systemOverlaysOnEnterFullScreen,
+    List<DeviceOrientation>? deviceOrientationsOnEnterFullScreen,
+    List<SystemUiOverlay>? systemOverlaysAfterFullScreen,
+    List<DeviceOrientation>? deviceOrientationsAfterFullScreen,
+    Duration? progressIndicatorDelay,
+    Widget Function(
+      BuildContext,
+      Animation<double>,
+      Animation<double>,
+      ChewieControllerProvider,
+    )? routePageBuilder,
+  }) {
+    return ChewieController(
+      draggableProgressBar: draggableProgressBar ?? this.draggableProgressBar,
+      videoPlayerController:
+          videoPlayerController ?? this.videoPlayerController,
+      optionsTranslation: optionsTranslation ?? this.optionsTranslation,
+      aspectRatio: aspectRatio ?? this.aspectRatio,
+      autoInitialize: autoInitialize ?? this.autoInitialize,
+      autoPlay: autoPlay ?? this.autoPlay,
+      startAt: startAt ?? this.startAt,
+      looping: looping ?? this.looping,
+      fullScreenByDefault: fullScreenByDefault ?? this.fullScreenByDefault,
+      cupertinoProgressColors:
+          cupertinoProgressColors ?? this.cupertinoProgressColors,
+      materialProgressColors:
+          materialProgressColors ?? this.materialProgressColors,
+      placeholder: placeholder ?? this.placeholder,
+      overlay: overlay ?? this.overlay,
+      showControlsOnInitialize:
+          showControlsOnInitialize ?? this.showControlsOnInitialize,
+      showOptions: showOptions ?? this.showOptions,
+      optionsBuilder: optionsBuilder ?? this.optionsBuilder,
+      additionalOptions: additionalOptions ?? this.additionalOptions,
+      showControls: showControls ?? this.showControls,
+      subtitle: subtitle ?? this.subtitle,
+      subtitleBuilder: subtitleBuilder ?? this.subtitleBuilder,
+      customControls: customControls ?? this.customControls,
+      errorBuilder: errorBuilder ?? this.errorBuilder,
+      allowedScreenSleep: allowedScreenSleep ?? this.allowedScreenSleep,
+      isLive: isLive ?? this.isLive,
+      allowFullScreen: allowFullScreen ?? this.allowFullScreen,
+      allowMuting: allowMuting ?? this.allowMuting,
+      allowPlaybackSpeedChanging:
+          allowPlaybackSpeedChanging ?? this.allowPlaybackSpeedChanging,
+      useRootNavigator: useRootNavigator ?? this.useRootNavigator,
+      playbackSpeeds: playbackSpeeds ?? this.playbackSpeeds,
+      systemOverlaysOnEnterFullScreen: systemOverlaysOnEnterFullScreen ??
+          this.systemOverlaysOnEnterFullScreen,
+      deviceOrientationsOnEnterFullScreen:
+          deviceOrientationsOnEnterFullScreen ??
+              this.deviceOrientationsOnEnterFullScreen,
+      systemOverlaysAfterFullScreen:
+          systemOverlaysAfterFullScreen ?? this.systemOverlaysAfterFullScreen,
+      deviceOrientationsAfterFullScreen: deviceOrientationsAfterFullScreen ??
+          this.deviceOrientationsAfterFullScreen,
+      routePageBuilder: routePageBuilder ?? this.routePageBuilder,
+      hideControlsTimer: hideControlsTimer ?? this.hideControlsTimer,
+      progressIndicatorDelay:
+          progressIndicatorDelay ?? this.progressIndicatorDelay,
+    );
+  }
+
+  static const defaultHideControlsTimer = Duration(seconds: 3);
+
+  /// If false, the options button in MaterialUI and MaterialDesktopUI
+  /// won't be shown.
+  final bool showOptions;
+
+  /// Pass your translations for the options like:
+  /// - PlaybackSpeed
+  /// - Subtitles
+  /// - Cancel
+  ///
+  /// Buttons
+  ///
+  /// These are required for the default `OptionItem`'s
+  final OptionsTranslation? optionsTranslation;
+
+  /// Build your own options with default chewieOptions shiped through
+  /// the builder method. Just add your own options to the Widget
+  /// you'll build. If you want to hide the chewieOptions, just leave them
+  /// out from your Widget.
+  final Future<void> Function(
+    BuildContext context,
+    List<OptionItem> chewieOptions,
+  )? optionsBuilder;
+
+  /// Add your own additional options on top of chewie options
+  final List<OptionItem> Function(BuildContext context)? additionalOptions;
+
+  /// Define here your own Widget on how your n'th subtitle will look like
+  Widget Function(BuildContext context, dynamic subtitle)? subtitleBuilder;
+
+  /// Add a List of Subtitles here in `Subtitles.subtitle`
+  Subtitles? subtitle;
 
   /// The controller for the video you want to play
   final VideoPlayerController videoPlayerController;
@@ -250,46 +434,59 @@ class ChewieController extends ChangeNotifier {
   /// Play the video as soon as it's displayed
   final bool autoPlay;
 
+  /// Non-Draggable Progress Bar
+  final bool draggableProgressBar;
+
   /// Start video at a certain position
-  final Duration startAt;
+  final Duration? startAt;
 
   /// Whether or not the video should loop
   final bool looping;
 
-  /// Weather or not to show the controls when initializing the widget.
+  /// Wether or not to show the controls when initializing the widget.
   final bool showControlsOnInitialize;
 
   /// Whether or not to show the controls at all
   final bool showControls;
 
+  /// Controller to pass into the [InteractiveViewer] component
+  final TransformationController? transformationController;
+
+  /// Whether or not to allow zooming and panning
+  final bool zoomAndPan;
+
+  /// Max scale when zooming
+  final double maxScale;
+
   /// Defines customised controls. Check [MaterialControls] or
   /// [CupertinoControls] for reference.
-  final Widget customControls;
+  final Widget? customControls;
 
-  /// When the video playback runs  into an error, you can build a custom
+  /// When the video playback runs into an error, you can build a custom
   /// error message.
-  final Widget Function(BuildContext context, String errorMessage) errorBuilder;
+  final Widget Function(BuildContext context, String errorMessage)?
+      errorBuilder;
 
   /// The Aspect Ratio of the Video. Important to get the correct size of the
   /// video!
   ///
   /// Will fallback to fitting within the space allowed.
-  final double aspectRatio;
+  final double? aspectRatio;
 
   /// The colors to use for controls on iOS. By default, the iOS player uses
   /// colors sampled from the original iOS 11 designs.
-  final ChewieProgressColors cupertinoProgressColors;
+  final ChewieProgressColors? cupertinoProgressColors;
 
   /// The colors to use for the Material Progress Bar. By default, the Material
   /// player uses the colors from your Theme.
-  final ChewieProgressColors materialProgressColors;
+  final ChewieProgressColors? materialProgressColors;
 
   /// The placeholder is displayed underneath the Video before it is initialized
   /// or played.
-  final Widget placeholder;
+  final Widget? placeholder;
 
   /// A widget which is placed between the video and the controls
-  final Widget overlay;
+  final Widget? overlay;
 
   /// Defines if the player will start in fullscreen when play is pressed
   final bool fullScreenByDefault;
@@ -297,7 +494,7 @@ class ChewieController extends ChangeNotifier {
   /// Defines if the player will sleep in fullscreen or not
   final bool allowedScreenSleep;
 
-  /// Defines if the controls should be for live stream video
+  /// Defines if the controls should be shown for live stream video
   final bool isLive;
 
   /// Defines if the fullscreen control should be shown
@@ -309,14 +506,20 @@ class ChewieController extends ChangeNotifier {
   /// Defines if the playback speed control should be shown
   final bool allowPlaybackSpeedChanging;
 
+  /// Defines if push/pop navigations use the rootNavigator
+  final bool useRootNavigator;
+
+  /// Defines the [Duration] before the video controls are hidden. By default, this is set to three seconds.
+  final Duration hideControlsTimer;
+
   /// Defines the set of allowed playback speeds user can change
   final List<double> playbackSpeeds;
 
   /// Defines the system overlays visible on entering fullscreen
-  final List<SystemUiOverlay> systemOverlaysOnEnterFullScreen;
+  final List<SystemUiOverlay>? systemOverlaysOnEnterFullScreen;
 
   /// Defines the set of allowed device orientations on entering fullscreen
-  final List<DeviceOrientation> deviceOrientationsOnEnterFullScreen;
+  final List<DeviceOrientation>? deviceOrientationsOnEnterFullScreen;
 
   /// Defines the system overlays visible after exiting fullscreen
   final List<SystemUiOverlay> systemOverlaysAfterFullScreen;
@@ -325,11 +528,18 @@ class ChewieController extends ChangeNotifier {
   final List<DeviceOrientation> deviceOrientationsAfterFullScreen;
 
   /// Defines a custom RoutePageBuilder for the fullscreen
-  final ChewieRoutePageBuilder routePageBuilder;
+  final ChewieRoutePageBuilder? routePageBuilder;
+
+  /// Defines a delay in milliseconds between entering buffering state and displaying the loading spinner. Set null (default) to disable it.
+  final Duration? progressIndicatorDelay;
+
+  /// Adds additional padding to the controls' [SafeArea] as desired.
+  /// Defaults to [EdgeInsets.zero].
+  final EdgeInsets controlsSafeAreaMinimum;
 
   static ChewieController of(BuildContext context) {
     final chewieControllerProvider =
-        context.dependOnInheritedWidgetOfExactType<_ChewieControllerProvider>();
+        context.dependOnInheritedWidgetOfExactType<ChewieControllerProvider>()!;
 
     return chewieControllerProvider.controller;
   }
@@ -340,11 +550,11 @@ class ChewieController extends ChangeNotifier {
 
   bool get isPlaying => videoPlayerController.value.isPlaying;
 
-  Future _initialize() async {
+  Future<dynamic> _initialize() async {
     await videoPlayerController.setLooping(looping);
 
     if ((autoInitialize || autoPlay) &&
-        !videoPlayerController.value.initialized) {
+        !videoPlayerController.value.isInitialized) {
       await videoPlayerController.initialize();
     }
 
@@ -357,7 +567,7 @@ class ChewieController extends ChangeNotifier {
     }
 
     if (startAt != null) {
-      await videoPlayerController.seekTo(startAt);
+      await videoPlayerController.seekTo(startAt!);
     }
 
     if (fullScreenByDefault) {
@@ -411,20 +621,22 @@ class ChewieController extends ChangeNotifier {
   Future<void> setVolume(double volume) async {
     await videoPlayerController.setVolume(volume);
   }
+
+  void setSubtitle(List<Subtitle> newSubtitle) {
+    subtitle = Subtitles(newSubtitle);
+  }
 }
 
-class _ChewieControllerProvider extends InheritedWidget {
-  const _ChewieControllerProvider({
-    Key key,
-    @required this.controller,
-    @required Widget child,
-  })  : assert(controller != null),
-        assert(child != null),
-        super(key: key, child: child);
+class ChewieControllerProvider extends InheritedWidget {
+  const ChewieControllerProvider({
+    Key? key,
+    required this.controller,
+    required Widget child,
+  }) : super(key: key, child: child);
 
   final ChewieController controller;
 
   @override
-  bool updateShouldNotify(_ChewieControllerProvider old) =>
-      controller != old.controller;
+  bool updateShouldNotify(ChewieControllerProvider oldWidget) =>
+      controller != oldWidget.controller;
 }
