@@ -1,20 +1,18 @@
-import 'dart:ui';
-
 import 'package:chewie/src/chewie_player.dart';
-import 'package:chewie/src/cupertino_controls.dart';
-import 'package:chewie/src/material_controls.dart';
-import 'package:flutter/foundation.dart';
+import 'package:chewie/src/helpers/adaptive_controls.dart';
+import 'package:chewie/src/notifiers/index.dart';
 import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
 import 'package:video_player/video_player.dart';
 
 class PlayerWithControls extends StatelessWidget {
-  const PlayerWithControls({Key key}) : super(key: key);
+  const PlayerWithControls({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     final ChewieController chewieController = ChewieController.of(context);
 
-    double _calculateAspectRatio(BuildContext context) {
+    double calculateAspectRatio(BuildContext context) {
       final size = MediaQuery.of(context).size;
       final width = size.width;
       final height = size.height;
@@ -22,39 +20,64 @@ class PlayerWithControls extends StatelessWidget {
       return width > height ? width / height : height / width;
     }
 
-    Widget _buildControls(
+    Widget buildControls(
       BuildContext context,
       ChewieController chewieController,
     ) {
-      final controls = Theme.of(context).platform == TargetPlatform.android
-          ? const MaterialControls()
-          : const CupertinoControls(
-              backgroundColor: Color.fromRGBO(41, 41, 41, 0.7),
-              iconColor: Color.fromARGB(255, 200, 200, 200),
-            );
       return chewieController.showControls
-          ? chewieController.customControls ?? controls
-          : Container();
+          ? chewieController.customControls ?? const AdaptiveControls()
+          : const SizedBox();
     }
 
-    Stack _buildPlayerWithControls(
-        ChewieController chewieController, BuildContext context) {
+    Widget buildPlayerWithControls(
+      ChewieController chewieController,
+      BuildContext context,
+    ) {
       return Stack(
         children: <Widget>[
-          chewieController.placeholder ?? Container(),
-          Center(
-            child: AspectRatio(
-              aspectRatio: chewieController.aspectRatio ??
-                  chewieController.videoPlayerController.value.aspectRatio,
-              child: VideoPlayer(chewieController.videoPlayerController),
+          if (chewieController.placeholder != null)
+            chewieController.placeholder!,
+          InteractiveViewer(
+            transformationController: chewieController.transformationController,
+            maxScale: chewieController.maxScale,
+            panEnabled: chewieController.zoomAndPan,
+            scaleEnabled: chewieController.zoomAndPan,
+            child: Center(
+              child: AspectRatio(
+                aspectRatio: chewieController.aspectRatio ??
+                    chewieController.videoPlayerController.value.aspectRatio,
+                child: VideoPlayer(chewieController.videoPlayerController),
+              ),
             ),
           ),
-          chewieController.overlay ?? Container(),
+          if (chewieController.overlay != null) chewieController.overlay!,
+          if (Theme.of(context).platform != TargetPlatform.iOS)
+            Consumer<PlayerNotifier>(
+              builder: (
+                BuildContext context,
+                PlayerNotifier notifier,
+                Widget? widget,
+              ) =>
+                  Visibility(
+                visible: !notifier.hideStuff,
+                child: AnimatedOpacity(
+                  opacity: notifier.hideStuff ? 0.0 : 0.8,
+                  duration: const Duration(
+                    milliseconds: 250,
+                  ),
+                  child: const DecoratedBox(
+                    decoration: BoxDecoration(color: Colors.black54),
+                    child: SizedBox.expand(),
+                  ),
+                ),
+              ),
+            ),
           if (!chewieController.isFullScreen)
-            _buildControls(context, chewieController)
+            buildControls(context, chewieController)
           else
             SafeArea(
-              child: _buildControls(context, chewieController),
+              bottom: false,
+              child: buildControls(context, chewieController),
             ),
         ],
       );
@@ -65,8 +88,8 @@ class PlayerWithControls extends StatelessWidget {
         height: MediaQuery.of(context).size.height,
         width: MediaQuery.of(context).size.width,
         child: AspectRatio(
-          aspectRatio: _calculateAspectRatio(context),
-          child: _buildPlayerWithControls(chewieController, context),
+          aspectRatio: calculateAspectRatio(context),
+          child: buildPlayerWithControls(chewieController, context),
         ),
       ),
     );
