@@ -2,5 +2,7 @@ library chewie;
 
 export 'src/chewie_player.dart';
 export 'src/chewie_progress_colors.dart';
-export 'src/cupertino_controls.dart';
-export 'src/material_controls.dart';
+export 'src/cupertino/cupertino_controls.dart';
+export 'src/material/material_controls.dart';
+export 'src/material/material_desktop_controls.dart';
+export 'src/models/index.dart';
